refactor(api-6l1q): relocate structure-entity services under services/structure/
