+++ apps/api/src/__tests__/services/structure/association.test.ts
@@ -1,15 +1,15 @@
 import { brandId } from "@pluralscape/types";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import { mockDb } from "../helpers/mock-db.js";
-import { mockOwnershipFailure } from "../helpers/mock-ownership.js";
-import { makeTestAuth } from "../helpers/test-auth.js";
+import { mockDb } from "../../helpers/mock-db.js";
+import { mockOwnershipFailure } from "../../helpers/mock-ownership.js";
+import { makeTestAuth } from "../../helpers/test-auth.js";
 
 import type { SystemId, SystemStructureEntityAssociationId } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────────
 
-vi.mock("../../lib/system-ownership.js", () => ({
+vi.mock("../../../lib/system-ownership.js", () => ({
   assertSystemOwnership: vi.fn(),
 }));
 
@@ -65,7 +65,7 @@ vi.mock("drizzle-orm", async (importOriginal) => {
 
 // ── Imports after mocks ──────────────────────────────────────────────
 
-const { assertSystemOwnership } = await import("../../lib/system-ownership.js");
+const { assertSystemOwnership } = await import("../../../lib/system-ownership.js");
 const { CreateStructureEntityAssociationBodySchema } = await import("@pluralscape/validation");
 
 const {
@@ -73,7 +73,7 @@ const {
   listEntityAssociations,
   deleteEntityAssociation,
   getEntityHierarchy,
-} = await import("../../services/structure-entity-association.service.js");
+} = await import("../../../services/structure/association.js");
 
 // ── Fixtures ─────────────────────────────────────────────────────────
