+++ apps/api/src/routes/structure/entities/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { getStructureEntity } from "../../../services/structure-entity.service.js";
+import { getStructureEntity } from "../../../services/structure/entity-crud/queries.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
