+++ apps/api/src/routes/structure/entities/hierarchy.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { getEntityHierarchy } from "../../../services/structure-entity.service.js";
+import { getEntityHierarchy } from "../../../services/structure/association.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
