+++ apps/api/src/routes/structure/entities/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { restoreStructureEntity } from "../../../services/structure-entity.service.js";
+import { restoreStructureEntity } from "../../../services/structure/entity-crud/lifecycle.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
