+++ apps/api/src/routes/structure/entities/update.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { updateStructureEntity } from "../../../services/structure-entity.service.js";
+import { updateStructureEntity } from "../../../services/structure/entity-crud/update.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
