+++ apps/api/src/routes/structure/entity-links/delete.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../../lib/audit-writer.js";
 import { getDb } from "../../../lib/db.js";
 import { requireIdParam } from "../../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { deleteEntityLink } from "../../../services/structure-entity.service.js";
+import { deleteEntityLink } from "../../../services/structure/link.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
