+++ apps/api/src/routes/structure/entity-links/update.ts
@@ -7,7 +7,7 @@ import { requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { updateEntityLink } from "../../../services/structure-entity.service.js";
+import { updateEntityLink } from "../../../services/structure/link.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
