+++ apps/api/src/routes/structure/entity-member-links/create.ts
@@ -9,7 +9,7 @@ import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createIdempotencyMiddleware } from "../../../middleware/idempotency.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { createEntityMemberLink } from "../../../services/structure-entity.service.js";
+import { createEntityMemberLink } from "../../../services/structure/member-link.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
