+++ apps/api/src/routes/structure/entity-member-links/delete.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../../lib/audit-writer.js";
 import { getDb } from "../../../lib/db.js";
 import { requireIdParam } from "../../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { deleteEntityMemberLink } from "../../../services/structure-entity.service.js";
+import { deleteEntityMemberLink } from "../../../services/structure/member-link.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
