+++ apps/api/src/routes/structure/entity-member-links/list.ts
@@ -6,7 +6,7 @@ import { requireIdParam } from "../../../lib/id-param.js";
 import { parseCursor, parsePaginationLimit } from "../../../lib/pagination.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
 import { DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT } from "../../../service.constants.js";
-import { listEntityMemberLinks } from "../../../services/structure-entity.service.js";
+import { listEntityMemberLinks } from "../../../services/structure/member-link.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
