+++ apps/api/src/routes/structure/entity-types/archive.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../../lib/audit-writer.js";
 import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { archiveEntityType } from "../../../services/structure-entity.service.js";
+import { archiveEntityType } from "../../../services/structure/entity-type/archive.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
