+++ apps/api/src/routes/structure/entity-types/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { getEntityType } from "../../../services/structure-entity.service.js";
+import { getEntityType } from "../../../services/structure/entity-type/get.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
