+++ apps/api/src/routes/structure/entity-types/list.ts
@@ -7,7 +7,7 @@ import { requireIdParam } from "../../../lib/id-param.js";
 import { parseCursor, parsePaginationLimit } from "../../../lib/pagination.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
 import { DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT } from "../../../service.constants.js";
-import { listEntityTypes } from "../../../services/structure-entity.service.js";
+import { listEntityTypes } from "../../../services/structure/entity-type/list.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
