+++ apps/api/src/routes/structure/entity-types/update.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../../lib/id-param.js";
 import { parseJsonBody } from "../../../lib/parse-json-body.js";
 import { envelope } from "../../../lib/response.js";
 import { createCategoryRateLimiter } from "../../../middleware/rate-limit.js";
-import { updateEntityType } from "../../../services/structure-entity.service.js";
+import { updateEntityType } from "../../../services/structure/entity-type/update.js";
 
 import type { AuthEnv } from "../../../lib/auth-context.js";
