--- apps/api/src/services/structure-entity.service.ts
@@ -1,16 +1,0 @@
-export * from "./structure/entity-crud/create.js";
-export * from "./structure/entity-crud/internal.js";
-export * from "./structure/entity-crud/lifecycle.js";
-export * from "./structure/entity-crud/queries.js";
-export * from "./structure/entity-crud/update.js";
-export * from "./structure/entity-type/create.js";
-export * from "./structure/entity-type/list.js";
-export * from "./structure/entity-type/get.js";
-export * from "./structure/entity-type/update.js";
-export * from "./structure/entity-type/archive.js";
-export * from "./structure/entity-type/restore.js";
-export * from "./structure/entity-type/delete.js";
-export type { EntityTypeResult } from "./structure/entity-type/internal.js";
-export * from "./structure-entity-link.service.js";
-export * from "./structure-entity-member-link.service.js";
-export * from "./structure-entity-association.service.js";
